refactor: Phase 3.1b — extract InputPanelMixin from ElectrodeAdvisorWidget
Move 9 input panel creation methods (351 lines) to InputPanelMixin
in main_window_input_panel.py. Combined with DrawingMixin, the
monolith has been reduced from 4,387 to 3,553 lines (-19.0%).

Methods extracted:
- _create_input_panel
- _create_electrical_measurements_section
- _create_power_factor_section
- _create_physical_parameters_section
- _create_electrode_depths_section
- _create_model_parameters_section
- _create_glass_properties_section
- _create_optimization_button
